refactor(ios-native): split app into PWA-style feature architecture
