+++ taskify-ios-native/Sources/TaskifyApp/App/TaskifyApp.swift
@@ -1,0 +1,14 @@
+import SwiftUI
+
+@main
+struct TaskifyApp: App {
+    @StateObject private var authVM = AppAuthViewModel()
+
+    var body: some Scene {
+        WindowGroup {
+            RootView()
+                .environmentObject(authVM)
+                .task { await authVM.bootstrap() }
+        }
+    }
+}
